+++ api-derive/imOnline/receivedHeartbeats.ts
@@ -2,14 +2,14 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import type { Observable } from 'https://esm.sh/rxjs@7.5.5';
-import type { Option, u32, WrapperOpaque } from 'https://deno.land/x/polkadot@0.0.0-8/types/mod.ts';
-import type { AccountId } from 'https://deno.land/x/polkadot@0.0.0-8/types/interfaces.ts';
-import type { PalletImOnlineBoundedOpaqueNetworkState } from 'https://deno.land/x/polkadot@0.0.0-8/types/lookup.ts';
+import type { Option, u32, WrapperOpaque } from 'https://deno.land/x/polkadot/types/mod.ts';
+import type { AccountId } from 'https://deno.land/x/polkadot/types/interfaces/index.ts';
+import type { PalletImOnlineBoundedOpaqueNetworkState } from 'https://deno.land/x/polkadot/types/lookup.ts';
 import type { DeriveApi, DeriveHeartbeats } from '../types.ts';
 
 import { combineLatest, map, of, switchMap } from 'https://esm.sh/rxjs@7.5.5';
 
-import { BN_ZERO } from 'https://deno.land/x/polkadot@0.0.0-8/util/mod.ts';
+import { BN_ZERO } from 'https://deno.land/x/polkadot/util/mod.ts';
 
 import { memo } from '../util/index.ts';