+++ api-derive/staking/stakerSlashes.ts
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import type { Observable } from 'https://esm.sh/rxjs@7.5.5';
-import type { EraIndex } from 'https://deno.land/x/polkadot@0.0.0-8/types/interfaces.ts';
+import type { EraIndex } from 'https://deno.land/x/polkadot/types/interfaces/index.ts';
 import type { DeriveApi, DeriveStakerSlashes } from '../types.ts';
 
 import { map } from 'https://esm.sh/rxjs@7.5.5';