+++ api-derive/staking/stashes.ts
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import type { Observable } from 'https://esm.sh/rxjs@7.5.5';
-import type { AccountId } from 'https://deno.land/x/polkadot@0.0.0-8/types/interfaces.ts';
+import type { AccountId } from 'https://deno.land/x/polkadot/types/interfaces/index.ts';
 import type { DeriveApi } from '../types.ts';
 
 import { map, startWith, switchMap } from 'https://esm.sh/rxjs@7.5.5';