+++ types-augment/lookup/types-kusama.ts
@@ -1,10 +1,10 @@
 // Auto-generated via `yarn polkadot-types-from-defs`, do not edit
 /* eslint-disable */
 
-declare module 'https://deno.land/x/polkadot@0.0.0-8/types/lookup.ts' {
-  import type { Compact, Enum, Null, Struct, Vec, u16, u32 } from 'https://deno.land/x/polkadot@0.0.0-8/types-codec/mod.ts';
-  import type { ITuple } from 'https://deno.land/x/polkadot@0.0.0-8/types-codec/types.ts';
-  import type { PerU16 } from 'https://deno.land/x/polkadot@0.0.0-8/types/interfaces/runtime.ts';
+declare module 'https://deno.land/x/polkadot/types/lookup.ts' {
+  import type { Compact, Enum, Null, Struct, Vec, u16, u32 } from 'https://deno.land/x/polkadot/types-codec/mod.ts';
+  import type { ITuple } from 'https://deno.land/x/polkadot/types-codec/types/index.ts';
+  import type { PerU16 } from 'https://deno.land/x/polkadot/types/interfaces/runtime/index.ts';
 
   /** @name KusamaRuntimeProxyType (79) */
   export interface KusamaRuntimeProxyType extends Enum {