+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## master
 
+- api, api-augment, api-base, api-contract, api-derive, rpc-augment, rpc-core, rpc-provider, types, types-augment, types-codec, types-create, types-known 8.9.2-19-x
 - wasm-bridge, wasm-crypto, wasm-crypto-asmjs, wasm-crypto-init, wasm-crypto-wasm, wasm-util 6.1.6-2-x
 - hw-ledger, hw-ledger-transports, keyring, networks, util, util-crypto, x-bigint, x-fetch, x-global, x-randomvalues, x-textdecoder, x-textencoder, x-ws 9.6.2-2-x
 - phishing 0.15.1